+++ geotrek/signage/templatetags/signage_tags.py
@@ -1,4 +1,5 @@
 from django import template
+from django.conf import settings
 from django.utils.translation import gettext_lazy as _
 from django.template.defaultfilters import stringfilter
 
@@ -18,3 +19,13 @@ def meters(value):
 @register.filter
 def class_name(value):
     return value.__class__.__name__
+
+
+@register.simple_tag
+def blades_enabled():
+    return settings.BLADE_ENABLED
+
+
+@register.simple_tag
+def lines_enabled():
+    return settings.LINE_ENABLED