+++ docs/changelog.rst
@@ -5,6 +5,11 @@ CHANGELOG
 2.72.0+dev (XXXX-XX-XX)
 -----------------------
 
+**New features**
+
+- Add BLADE_ENABLED setting to hide Blade in Signage forms and in Signage detail page
+- Add LINE_ENABLED setting to hide Line in Blade forms and in Blade detail page
+
 **User interface**
 
 - Clarify Land Edge module browsing (#1404)
@@ -16,7 +21,6 @@ CHANGELOG
 - Merge tabs in Intervention forms (#2712)
 - Make targets display more specific in Interventions exports (#2711)
 - Improve support for Tourinsoft v3 with new medias management
-
 
 
 2.72.0 (2021-11-16)